+++ pgscatalog.core/src/pgscatalog/core/lib/__init__.py
@@ -1,8 +1,11 @@
 from ._config import Config
+from .allele import Allele
+from .effecttype import EffectType
+from .genomebuild import GenomeBuild
 from .catalogapi import ScoreQueryResult, CatalogQuery, CatalogCategory
 from .scorefiles import ScoringFiles, ScoringFile, NormalisedScoringFile
-from .scorevariant import ScoreVariant, EffectType
-from .genomebuild import GenomeBuild
+from .catalogscorevariant import CatalogScoreVariant
+from .scorevariant import ScoreVariant
 from .targetvariants import TargetVariants, TargetVariant, TargetType
 from ._relabel import RelabelArgs, relabel, relabel_write
 from ._sortpaths import effect_type_keyfunc, chrom_keyfunc
@@ -56,10 +59,12 @@ __all__ = [
     "TargetVariants",
     "TargetType",
     "NormalisedScoringFile",
-    "EffectType",
     "RelabelArgs",
     "relabel",
     "relabel_write",
     "effect_type_keyfunc",
     "chrom_keyfunc",
+    "CatalogScoreVariant",
+    "EffectType",
+    "Allele",
 ]
